+++ CineHive/src/main/java/com/example/CineHive/controller/myPage/MyPageController.java
@@ -1,31 +1,22 @@
 package com.example.CineHive.controller.myPage;
 
 import com.example.CineHive.dto.board.BoardDto;
-import com.example.CineHive.dto.board.GetListBoardDto;
 import com.example.CineHive.dto.comment.CommentDto;
-import com.example.CineHive.dto.user.ChangeMemNameRequest;
-import com.example.CineHive.dto.user.ChangeMemSexRequest;
-import com.example.CineHive.dto.user.ChangePasswordRequest;
 import com.example.CineHive.entity.User;
 import com.example.CineHive.entity.board.*;
 
 import com.example.CineHive.entity.reply.Reply;
 import com.example.CineHive.entity.videotype.Movie;
-import com.example.CineHive.mapper.CommentMapper;
-import com.example.CineHive.repository.LoginHistoryRepository;
 import com.example.CineHive.repository.UserRepository;
 import com.example.CineHive.repository.board.*;
 import com.example.CineHive.repository.reply.ReplyRepository;
 import com.example.CineHive.repository.videos.movie.MovieRepository;
-import com.example.CineHive.service.UserService;
-import com.example.CineHive.service.board.BoardService;
 import com.example.CineHive.service.reply.ReplyBookmarkService;
 import com.example.CineHive.util.JwtTokenUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -47,7 +38,6 @@ public class MyPageController {
     private final MovieRepository movieRepository;
     private final BoardRepository boardRepository;
     private final ReplyRepository replyRepository;
-    private final UserService userService;
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
     private final LikeRepository likeRepository;
@@ -350,77 +340,6 @@ public class MyPageController {
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("유효하지 않은 토큰입니다.");
-        }
-    }
-
-
-
-
-    @PutMapping("/change-password")
-    @Operation(summary = "비밀번호 변경", description = "기존 비밀번호 확인 후 새 비밀번호로 변경")
-    public ResponseEntity<?> changePassword(@RequestBody ChangePasswordRequest request, HttpServletRequest httpRequest) {
-        String token = jwtTokenUtil.extractTokenFromRequest(httpRequest);
-        if (token == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰이 필요합니다.");
-        }
-
-        try {
-            String email = jwtTokenUtil.extractUsername(token);
-            boolean result = userService.changePassword(email, request.getOldPassword(), request.getNewPassword());
-            if (result) {
-                return ResponseEntity.ok("비밀번호가 성공적으로 변경되었습니다.");
-            } else {
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("기존 비밀번호가 일치하지 않습니다.");
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("비밀번호 변경 중 오류 발생");
-        }
-    }
-
-
-    @PutMapping("/change-memname")
-    @Operation(summary = "이름 변경", description = "사용자가 입력한 이름으로 변경")
-    public ResponseEntity<?> changeMemName(@RequestBody ChangeMemNameRequest request, HttpServletRequest httpRequest) {
-        String token = jwtTokenUtil.extractTokenFromRequest(httpRequest);
-        if (token == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰이 필요합니다.");
-        }
-
-        try {
-            String email = jwtTokenUtil.extractUsername(token);
-            boolean result = userService.changeMemName(email, request.getNewMemName());
-            if (result) {
-                return ResponseEntity.ok("이름이 성공적으로 변경되었습니다.");
-            } else {
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("이름 변경 실패");
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("이름 변경 중 오류 발생");
-        }
-    }
-
-
-    @PutMapping("/change-memsex")
-    @Operation(summary = "성별 변경", description = "사용자가 입력한 성별로 변경 (male, female, other만 허용)")
-    public ResponseEntity<?> changeMemSex(@RequestBody ChangeMemSexRequest request, HttpServletRequest httpRequest) {
-        String token = jwtTokenUtil.extractTokenFromRequest(httpRequest);
-        if (token == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰이 필요합니다.");
-        }
-
-        try {
-            String email = jwtTokenUtil.extractUsername(token);
-            boolean result = userService.changeMemSex(email, request.getNewMemSex());
-            if (result) {
-                return ResponseEntity.ok("성별이 성공적으로 변경되었습니다.");
-            } else {
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("유효한 성별 값은 male, female, other 입니다.");
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("성별 변경 중 오류 발생");
         }
     }
 }
